pyproject.toml
[project]
name = "calsync"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click",
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth-httplib2",
]

[project.scripts]
calsync = "calsync.cli:main"
